[Gif Feed] Fix Text Input
Since the query was determined by app state, and we wait for the network
request to come back before emitting a new one, there were cases where
the searchQuery that was trying the input field would be slow.

The solution is to initially set the query via the AppState, but then
allow subsequent updates to update local state as well as shoot off an
action to update the AppState.